requirements.txt
flask
gunicorn
werkzeug
numpy
tensorflow
Pillow
